fix output transform

+++ kaggle_imsegm/transform.py
@@ -4,6 +4,7 @@ from typing import Any, Callable, List, Tuple, Union
 import albumentations as alb
 
 import numpy as np
+import torch
 from albumentations import Compose
 from albumentations.pytorch import ToTensorV2
 
@@ -152,11 +153,11 @@ def default_uncollate(batch: Any) -> List[Any]:
 class SemanticSegmentationOutputTransform(OutputTransform):
     def per_sample_transform(self, sample: Any) -> Any:
         resize = alb.Resize(*[s.item() for s in sample["metadata"]["size"]])
-        sample["input"] = sample["input"].cpu().numpy()
+        sample["input"] = sample["input"].to(torch.float32).cpu().numpy()
         if sample["input"].ndim == 3:
             sample["input"] = np.rollaxis(sample["input"], 0, 3)
         sample["input"] = resize(image=sample["input"])["image"]
-        sample["preds"] = [resize(image=pred)["image"] for pred in sample["preds"].cpu().numpy()]
+        sample["preds"] = [resize(image=pred)["image"] for pred in sample["preds"].to(torch.float32).cpu().numpy()]
         return super().per_sample_transform(sample)
 
     @staticmethod
